ZCOVetcher: move definitions from .h -> .cpp
Remotely related to https://github.com/openvstorage/volumedriver/issues/213
