+++ src/volumedriver/ZCOVetcher.h
@@ -16,163 +16,38 @@
 #ifndef ZCO_VETCHER_H
 #define ZCO_VETCHER_H
 
-#include "FailOverCacheConfigWrapper.h"
-#include "SCOAccessData.h"
-#include "SCOCache.h"
-#include "VolManager.h"
-#include "VolumeConfig.h"
+#include "Types.h"
 
 #include <memory>
 
-#include <backend/BackendInterface.h>
+#include <boost/filesystem/path.hpp>
+
+#include <youtils/IncrementalChecksum.h>
+#include <youtils/Logger.h>
+
+#include <backend/Namespace.h>
 
 namespace volumedriver
 {
 
+class SCOCache;
+class FailOverCacheConfigWrapper;
+class VolumeConfig;
+
 class ZCOVetcher
 {
 public:
-    ZCOVetcher(const VolumeConfig& volume_config)
-        : ns_(volume_config.getNS())
-        , sco_cache_(*VolManager::get()->getSCOCache())
-        , scosize_(volume_config.getSCOSize())
-        , lba_size_(volume_config.lba_size_)
-        , cluster_mult_(volume_config.cluster_mult_)
-    {
-        if(not sco_cache_.hasDisabledNamespace(ns_))
-        {
-            if(sco_cache_.hasNamespace(ns_))
-            {
-                LOG_WARN("Trying to restart from a namespace that is still active " << ns_);
-                throw fungi::IOException("ZCOVetcher: namespace still active");
-            }
-            else
-            {
-                LOG_WARN("Namespace " << ns_ <<
-                         " not found in scocache, local restart not possible");
-                throw fungi::IOException("ZCOVetcher: No such namespace");
-            }
-        }
-        BackendInterfacePtr bi(VolManager::get()->createBackendInterface(ns_));
-        SCOAccessDataPersistor sadp(bi->clone());
-        SCOAccessDataPtr sad(sadp.pull());
+    explicit ZCOVetcher(const VolumeConfig&);
 
-        foc_config_wrapper_ =
-            VolumeFactory::get_config_from_backend<FailOverCacheConfigWrapper>(*bi);
-        const uint64_t max_non_disposable =
-            VolManager::get()->get_sco_cache_max_non_disposable_bytes(volume_config);
-        sco_cache_.enableNamespace(ns_,
-                                   0,
-                                   max_non_disposable,
-                                   *sad);
-    }
-
-    ~ZCOVetcher()
-    {
-        sco_cache_.disableNamespace(ns_);
-    }
+    ~ZCOVetcher();
 
     CachedSCOPtr
-    getSCO(SCO sco)
-    {
-        if (foc_config_wrapper_->config())
-        {
-            LOG_INFO("Trying to get SCO " << ns_ << "/" << sco <<
-                     " from the DTL or the backend");
-
-            RawFailOverCacheSCOFetcher
-                fetcher(sco,
-                        *foc_config_wrapper_->config(),
-                        ns_,
-                        lba_size_,
-                        cluster_mult_,
-                        VolManager::get()->dtl_request_timeout(),
-                        VolManager::get()->dtl_connect_timeout());
-
-            return sco_cache_.getSCO(ns_,
-                                     sco,
-                                     scosize_,
-                                     fetcher);
-        }
-        else
-        {
-            LOG_INFO("No FailOverCache, trying to get SCO " << sco <<
-                     " from the backend");
-
-            return sco_cache_.findSCO(ns_,
-                                      sco);
-        }
-    }
+    getSCO(SCO sco);
 
     bool
-    checkSCO(ClusterLocation cluster_location,
-             CheckSum::value_type cs,
-             bool retry)
-    {
-        try
-        {
-            CachedSCOPtr sco_ptr = getSCO(cluster_location.sco());
-            if (not sco_ptr)
-            {
-                return false;
-            }
-
-            fs::path sco_path = sco_ptr->path();
-            if(sco_path != incremental_checksum.first)
-            {
-                incremental_checksum.second.reset(new IncrementalChecksum(sco_path));
-                incremental_checksum.first = sco_path;
-            }
-
-            try
-            {
-                const CheckSum& s =
-                    incremental_checksum.second->checksum((cluster_location.offset() + 1) *
-                                                          lba_size_ * cluster_mult_);
-
-                if(s.getValue() == cs)
-                {
-                    return true;
-                }
-            }
-            catch(youtils::CheckSumFileNotLargeEnough& e)
-            {
-                LOG_ERROR("File Not Large Enough");
-            }
-
-            if(retry and
-               foc_config_wrapper_->config())
-            {
-                fs::path tmp = FileUtils::create_temp_file(sco_path);
-                FileUtils::safe_copy(sco_path, tmp);
-                fs::remove(sco_path);
-                incremental_checksum.first.clear();
-                incremental_checksum.second.reset(0);
-
-                sco_cache_.removeSCO(ns_,
-                                     cluster_location.sco(),
-                                     true,
-                                     false);
-                if(checkSCO(cluster_location,
-                            cs,
-                            false))
-                {
-                    fs::remove(tmp);
-                    return true;
-                }
-
-                fs::remove(sco_path);
-                fs::rename(tmp, sco_path);
-            }
-
-            return false;
-        }
-        CATCH_STD_ALL_EWHAT({
-                LOG_ERROR("Could not retrieve SCO for clusterlocation " <<
-                          cluster_location << ": " << EWHAT);
-                return false;
-            });
-    }
+    checkSCO(ClusterLocation,
+             CheckSum::value_type,
+             bool retry);
 
 private:
     DECLARE_LOGGER("ZCOVetcher");
@@ -182,7 +57,7 @@ private:
     const uint64_t scosize_;
     const LBASize lba_size_;
     const ClusterMultiplier cluster_mult_;
-    std::pair<fs::path, std::unique_ptr<youtils::IncrementalChecksum> > incremental_checksum;
+    std::pair<boost::filesystem::path, std::unique_ptr<youtils::IncrementalChecksum> > incremental_checksum_;
     std::unique_ptr<FailOverCacheConfigWrapper> foc_config_wrapper_;
 };
